feat: add UCD path utilities and tests for Unicode version handling
